+++ FirebaseMessaging/CHANGELOG.md
@@ -1,9 +1,10 @@
 # 11.0.0
 - [fixed] Completed Messaging's transition to NSSecureCoding (#12343).
 
-# 10.29.0
+# Unreleased
 - [fixed] Renamed "initWithFileName" internal method that was causing submission issues for some
   users. (#13134).
+- [fixed] Fixed the APS Environment key on visionOS. (#13173)
 
 # 10.27.0
 - [fixed] Fixed bug preventing Messaging from working with a custom sqlite3