+++ FirebaseMessaging/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 10.29.0
 - [fixed] Renamed "initWithFileName" internal method that was causing submission issues for some
   users. (#13134).
 - [fixed] Fixed the APS Environment key on visionOS. (#13173)